move spinner
